Change type of entity class.

+++ C#/Csv/src/PersonEntity.cs
@@ -4,6 +4,6 @@ public class PersonEntity
 {
     public string Name { get; set; } = null!;
     public int Age { get; set; }
-    public int Height { get; set; }
-    public int Weight { get; set; }
+    public decimal Height { get; set; }
+    public decimal Weight { get; set; }
 }
